Put STLForecaster into trend.py
