+++ docs/source/api_reference/forecasting.rst
@@ -44,6 +44,7 @@ Trend
 
     TrendForecaster
     PolynomialTrendForecaster
+    STLForecaster
 
 Exponential Smoothing
 ---------------------
@@ -163,7 +164,6 @@ Composition
     DirRecTabularRegressionForecaster
     DirRecTimeSeriesRegressionForecaster
     MultiplexForecaster
-    STLForecaster
 
 .. autosummary::
     :toctree: auto_generated/
